Add files via upload
This is the current LED Control webpage code. Its VERY basic for now as I'm concentrating on the Arduino Sketch file 1st. This file MUST be located inside a folder called 'data' as a sub-directory of your Arduino Project Folder. This is where the ESP32 sketch data upload plugin expects to find it. The ESP32 sketch data upload plugin will then upload it to the appropriate directory on your ESP32.